Added two tools for parsing ram and sfr info
parse-gprbanks.lua is useful for figuring out whether a chip's ram
region contains any *holes*; it reports both the declared ram size
(RAMSIZE in the .ini file) and the contiugous ram space it finds in
GPRBANKS.

ram-info.lua is similar. It prints out the start of ram, the declared
size (RAMSIZE again), the number of bits in the BSR, and the start and
end addreses of the SFRs. The idea is to figure out how to safely zero
the ram at reset time without clobbering SFRs.

These are exploratory tools; the useful findings have been incorporated
into ini2mu.lua (a commit is coming very soon. ;-)